Clean up discovery scaffolding
