+++ src/resolve/tests/filter_tests.rs
@@ -30,7 +30,6 @@ fn include_filter_seeds_bootstrap_docs_without_hooks_or_mcp() {
                     name: item_name,
                 },
                 source_path,
-                layer: 0,
             },
         );
     }
